requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-api-python-client==2.120.0
python-dotenv==1.0.1
certifi==2024.2.2 